21-03-2023-2

+++ DSA/code_ninja/time&space_complexity/getTime.cpp
@@ -6,7 +6,7 @@
 #include "getTime.h"
 
 
- long getTimeInusec()
+  long long getTimeInusec()
 {
 	struct timeval current_time;
 	gettimeofday(&current_time, NULL);
