+++ DSA/code_ninja/time&space_complexity/getTime.h
@@ -2,6 +2,6 @@
 #ifndef __TIME_H
 #define __TIME_H
 
- long getTimeInusec();
+ long long getTimeInusec();
 
 #endif
